requirements.txt
attrs==21.4.0
colorama==0.4.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
tomli==2.0.1
